Django/requirements.txt
mysqlclient
django-allauth
python-dotenv